keep check for tmpfile

+++ preload_data.sh
@@ -218,7 +218,9 @@ function pre_req_bpm() {
 function cleanup() {
   title "Cleaning up any previous copy operations..."
   msg "-----------------------------------------------------------------------"
-  rm $TEMPFILE
+  if [[ -f $TEMPFILE ]]; then
+    rm $TEMPFILE
+  fi
   ${OC} delete job mongodb-backup -n $FROM_NAMESPACE
   ${OC} delete job mongodb-restore -n $TO_NAMESPACE
   pvcexists=$(${OC} get pvc cs-mongodump -n $FROM_NAMESPACE --no-headers --ignore-not-found | awk '{print $2}')
